Written by: David Lebel <lebel@openbsd.org>
initial import of gle-3.0.4:
--
The GLE Tubing and Extrusion Library is a graphics application
programming interface (API). The library consists of a number of
"C" language subroutines for drawing tubing and extrusions. It is
a very fast implementation of these shapes, outperforming all other
implementations, most by orders of magnitude.  The library is
distributed in source code form, in a package that includes
documentation, a VRML proposal, Makefiles, and full source code and
header files. It uses the OpenGL (TM) programming API to perform
the actual drawing of the tubing and extrusions.
 
GLE is designed for and maintained on the Linux operating system,
and is known to run on other Unix operating systems such as AIX,
IRIX, Ultrix and HPUX with OpenGL or Mesa. GLE is also known to run
on IBM OS/2 Warp, Apple Macintosh OS9, and Microsoft Windows NT and
has been used to develop screen-savers for some of these popular
PC operating systems.
 
WWW: http://www.linas.org/gle/