Written by: Peter Valchev <pvalchev@openbsd.org>
Update to nap-1.4.5
After more than 2 months, this is an overdue release.  There are a lot
of bugfixes and new features, see the distributed ChangeLog for details.

Main changes:
* The main new feature is that the OpenNap servers are now the default
for nap.  The reason for this is that Napster Inc. has modified their
servers to use a new client auth scheme, which means that nap users
can no longer connect to their servers.  Nap will automatically
download a list of available servers from napigator.com, and then
connect to the first available one.
* Nap's sourceforge page has been updated and now fully operational.
You are encouraged to use sourceforge's facilities for bug reports,
feature requests, etc.