Written by: David Lebel <lebel@openbsd.org>
initial import of sash-3.4:
--
SASH - Stand-alone shell with built-in commands.

The sash program is a stand-alone shell which is useful for recovering
from certain types of system failures. In particular, it was created
in order to cope with the problem of missing shared libraries or
important executable.

WWW: http://www.canb.auug.org.au/~dbell/

Submitted by Albert Chang <vedge@vedge.com.ar>