Written by: Peter Valchev <pvalchev@openbsd.org>
Import of grace-5.1.4

Grace is a tool to make two-dimensional plots of numerical data. It
runs under various (if not all) flavours of UNIX with X11 and M*tif.
Its capabilities are roughly similar to GUI-based programs like
Sigmaplot or Microcal Origin plus script-based tools like gnuplot or
Genplot. Its strength lies in the fact that it combines the
convenience of a graphical user interface with the power of a
scripting language which enables it to do sophisticated calculations
or perform automated tasks.

Flavors:
    no_x11 - build command line tool only, no gui
    fftw   - use math/fftw for optimized FFT
    netcdf - use math/netcdf for network common data form support
    xmhtml - use www/xmhtml to use grace's internal HTML browser
             for helpfiles etc.

WWW: http://plasma-gate.weizmann.ac.il/Grace/

MAINTAINER=	Nikolay Sturm <Nikolay.Sturm@desy.de>