Written by: David Lebel <lebel@openbsd.org>
initial import of sdlroids-1.3.4:
--
SDLRoids is essentially an Asteroids clone, but with a few extra
features, and some nice game physics. It is based of xhyperoid,
which is a UNIX port of the 16-bit Windows game Hyperoid. Major
changes from xhyperoid are that it's using SDL for sound and graphics,
has a couple of extra powerups and and that the shield behaves
differently.

The game can either run in windowed mode or in fullscreen mode. You
can (usually) switch between them during play by pressing Alt-Enter.

The play area (apart from the score/lives/etc. display) is square
and centred, just like in the the original Hyperoid did. This might
not might not change, bug for now that is how it works.

WWW: http://david.hedbor.org/projects/sdlroids/