Written by: Shell Hung <shell@openbsd.org>
Initial import of py-mxDateTime-2.0.2

Port by Maurice Nonnekes <maurice@amaze.nl>

---
The eGenix mx Extension Series are a collection of Python extensions
written in ANSI C and Python which provide a large spectrum of useful
additions to everyday Python programming.

The BASE package includes the Open Source subpackages of the series and
is needed by all other add-on packages of the series.