Written by: Shell Hung <shell@openbsd.org>
Initial import of Net-IPv4Addr-0.10

---

Net::IPv4Addr provides functions for parsing IPv4 addresses both
in traditional address/netmask format and in the new CIDR format.
There are also methods for calculating the network and broadcast
address and also to see check if a given address is in a specific
network.