Written by: Shell Hung <shell@openbsd.org>
Initial import of gpgme-0.2.2

Maintainer : COUDERC Damien <couderc.damien@wanadoo.fr>
---

GnuPG Made Easy (GPGME) is a library designed to make access to
GnuPG easier for applications.

It provides a High-Level Crypto API for encryption, decryption,
signing, signature verification and key management.  Currently it
uses GnuPG as it's backend but the API isn't restricted to this
engine; in fact it is planned to add other backends to it.