Written by: Brad Smith <brad@openbsd.org>
add http_load port
--
http_load runs multiple HTTP fetches in parallel, to test the throughput
of a web server.

However unlike most such test clients, it runs in a single process, so it
doesn't bog down the client machine. It can be configured to do HTTPS
fetches as well.