Written by: Peter Valchev <pvalchev@openbsd.org>
import of burgerspace-1.5

Burgerspace is a BurgerTime clone for X. You are a chef and must walk
over hamburger ingredients (buns, meat and lettuce) to make them fall
from floor to floor, until they end up in the plates at the bottom of
the screen.

Uses the gengameng library by the same author, which uses SDL itself.