Written by: Anil Madhavapeddy <avsm@openbsd.org>
initial import of wv-0.6.5
--

wv is a library which allows access to Microsoft Word files. It can
load and parse the word 2000, 97, 95 and 6 file formats. These are
the file formats known internally as Word 9, 8, 7 and 6. Work is
underway to support reading earlier formats as well: Word 2 docs
are converted to plaintext.