Written by: David Lebel <lebel@openbsd.org>
modify angband in order to use /var/games/zangband for
user modifiable files (like score, bone and save files). Add
pkg/INSTALL and pkg/DEINSTALL to manage them.

Also, sync the installation with the way zangband does it.

Instead of "No", specify that the license is unclear for distribution
rights on CDROM.