Written by: Christian Weisgerber <naddy@openbsd.org>
Import smtpclient 1.0.0.
Submitted by Patroklos Argyroudis <argp@bsd.gr>.

This program is a minimal SMTP client that takes an email
message body and passes it on to a SMTP server (default is the
MTA on the local host). Since it is completely self-supporting,
it is especially suitable for use in restricted environments.