Written by: David Lebel <lebel@openbsd.org>
Removed patch-g10_mainproc.c mostly because this patch was related to
the 1.0.4 release and because the problem was later fixed in 1.0.5 (and
later versions).  Here is a quote from the NEWS file about this issue:

--cut--
* WARNING: The semantics of --verify have changed to address a
problem with detached signature detection. --verify now ignores signed
material given on stdin unless this is requested by using a "-" as the
name for the file with the signed material.  Please check all your
detached signature handling applications and make sure that they don't
pipe the signed material to stdin without using a filename together with
"-" on the the command line.
--cut--

The patch introduce a problem in the way verifying a signature returned
a value of 2 instead of 0 (when the signature was properly verified).
The symptom showed itself in mutt for example.

This problem was found by Anders Arnholm <anders@arnholm.nu>

markus@ asked me to commit this.