Written by: Marc Espie <espie@openbsd.org>
Install the whole thing, including the shared library.
Make sure it's built properly by USE_LIBTOOL.
Use the default install target, just tweak the html files location.

Bump the minor number to leave room for a stub library to put on CDs
to allow for things to depend on lame.