Written by: Marc Espie <espie@openbsd.org>
lame support
Somewhat sneaky: link against a `stub' lame library so that the dependency
is a build dependency.

This allows us to put a lame-enabled kdebase on the CD: it will refer to
the stub library by default. When you grab a newer lame, it will have
a higher minor number, and be used instead (this is the old ssl/crypto
trick).