Written by: Peter Stromberg <wilfried@openbsd.org>
This is a system monitoring dockapp, visually based on the GNOME
"BubbleMon" applet (here). Basically, it displays CPU and memory
load as bubbles in a jar of water. But that's where similarity ends.
New bubblemon-dockapp features translucent CPU load meter (for
accurate CPU load measurement), yellow duck swimming back and forth
on the water surface (just for fun), and fading load average and
memory usage screens.