Written by: David Lebel <lebel@openbsd.org>
bump package revision version since we added four new official patches
to slrn version 0.9.7.2:
--
Bug fixes:

* slrn enters all groups when read_active is on
Fixed 2001-08-27

When read_active is on, slrn still enters all groups to find the
number of unread articles, which slows down startup considerably.

* force_authentication is ignored when re-connecting
Fixed 2001-08-27

When re-connecting, slrn only sends authentication data if the server
requests it, even if force_authentication is set.

* Request for authentication is not always recognized
Fixed 2001-08-30

In some rare cases, slrn does not notice that it needs to send
authentication data, although the server requests it. This can even
damage your newsrc file.

* Parse "news:" URLs enclosed in angle brackets correctly
Fixed 2001-08-30

When a "news:" URL is enclosed in angle brackets, slrn regards the
closing bracket as a part of it.