Written by: Marc Espie <espie@openbsd.org>
Use cookies for dependencies as well, and for the creation of a workdir.
This can speed up some iterations, as a dependency won't be checked more
than once. This also allows for make extract NO_DEPENDS=Yes, then later
to have the dependencies pulled in at configure stage (even though this
might not work all the time, as some dependencies may be needed earlier
but then you'll certainly notice).