Written by: Christian Weisgerber <naddy@openbsd.org>
Update to 16.1e:
Added code in both the client and the server to detect whether the
peer is an old version with the S1G bug.  The server will refuse
to serve such clients, and the client will refuse updates from
such a server.  In each case, an error message is printed with a
URL that describes the bug and the upgrade procedure.