Written by: Marc Espie <espie@openbsd.org>
Fix that as a conflict between libgtop-*-!gnome and gnome-libs-*
(note that we can't write libgtop-*-!gnome as a pkgcfl, but it doesn't
matter, because the check is done at gnome-libs-* install time, at which
point libgtop-*-gnome is definitely not going to be around yet).

Thanks to naddy@ for setting me straight.