Written by: Kevin Lo <kevlo@openbsd.org>
Initial import of kslide 0.83.

KSlide is a simple puzzle game compatible with the K Desktop Environment.
The playing principle of KSlide is very simple. From a square set of tiles,
one is removed, allowing at least two adjacent tiles to be shifted.
This way, the puzzle is scrambled by the program, and the player's goal
is to restore the original order.