Written by: David Lebel <lebel@openbsd.org>
initial import of net/ipband 0.6:
--
ipband is a pcap based IP traffic monitor. It tallies per-subnet traffic
and bandwidth usage and starts detailed logging if specified threshold
for the specific subnet is exceeded. If traffic has been high for a
certain period of time, the report for that subnet is generated which
can be appended to a file or e-mailed. When bandwidth usage drops below
the threshold, detailed logging for the subnet is stopped and memory is
freed.

This utility could be handy in a limited bandwidth WAN environment
(frame relay, ISDN etc. circuits) to pinpoint offending traffic source
if certain links become saturated to the point where legitimate packets
start getting dropped.

It also can be used to monitor internet connection when specifying the
range of local ip addresses (to avoid firing reports about non-local
networks).

WWW: http://ipband.sourceforge.net/

Submitted by Vincent Derrien <hyzzod@free.fr>