Written by: Reinhard J. Sammer <reinhard@openbsd.org>
makeself is a small shell script that generates a self-extractible
tar.gz archive from a directory. The resulting file appears as a
shell script, and can be launched as is. The archive will then
uncompress itself to a temporary directory and an arbitrary command
will be executed (for example, an installation script).