Written by: Peter Galbavy <peter@openbsd.org>
From the ChangeLog:

Version 3.34
------------
1. Exim was failing to diagnose a lone \ at the end of an expansion string as
an error (basically a typo in the code).

2. If logging was only to syslog, and Exim was trying to panic-die, it crashed
instead of dying cleanly.

3. If an address was routed using a DNS lookup that found no MX records, but
one or more A records, and fallback hosts were specified on the transport, the
fallback hosts were ignored.

4. $message_body_size was set incorrectly (to zero) during filter testing.

5. Ensure the configuration file is closed before running the -bi command.

6. Reap all complete processes within the loop for accepting -bs or -bS
messages, because it seems that not all OS do this automatically when SIGCHLD
is set to SIG_IGN.

7. Reset SIGHUP to SIG_IGN before restarting a daemon, in case another SIGHUP
arrives very quickly and kills the newly started Exim before it has a chance to
get going.

8. After "452 space shortage", was not unsetting the sender address. Could lead
to strange effects when the client was pipelining.

9. There was no check that getpeername() was giving a socket address when
called on stdin passed from a previous delivery.

10. If a local part beginning with a pipe symbol was directed to a pipe
transport, the transport got confused as to which command it should run.
This could be a security exposure if unchecked local parts are directed
or routed to pipe transports.