Written by: Mark Grimes <mark@openbsd.org>
Emacs 21.1.1 adds tty font-locking that does not play nice with termcap.
This bug is being looked into.  It only affects tty font-lock settings
within a .emacs file.  If you do not use color syntaxing in Emacs under
an xterm, this will not affect you.  For those that do, setting TERM
to xterm (rather then one of the color xterm variants) is a temporary
fix.