Written by: Matt Behrens <matt@openbsd.org>
Restore shared libpython.  The problem was that pthreads' 64k stack
size is not adequate for some of the complex processing that Zope
does, and -fPIC was aggravating that condition.  Now, if we're using
shared libs, up the stack size to 128k.

Also remove some stray files that never should have been committed
in the first place.