Written by: Matt Behrens <matt@openbsd.org>
Upgrade to 2.5.1b1, as it's got a few important fixes, and nobody
knows if 2.5.1 is going to come out in time for the release.

Also add my patch to fix setuid support (not used by default) and
set our own version string to distinguish this from an "official"
Zope release.