Written by: Marc Espie <espie@openbsd.org>
Let make update-plist look at dependent packages (LIB_DEPENDS and
RUN_DEPENDS) to create a more accurate list of @dirrm.

Minor negative side-effect: up-to-date dependent packages must be present
for this to work. Not a problem, since this is only used by maintainers.

This improves the quality of generated plists for large packages with lots
of dependencies by a large amount (e.g., kde)

reviewed by pval@