Written by: Marc Espie <espie@openbsd.org>
apparently, linux expect ld.so to be able to fill back references to
program code from dynamic libraries.

This does not work on our ELF platforms. Just put the getmntent code inside
the one source file that needs it, as a static function so that it does not
show outside the library.