Written by: Marc Espie <espie@openbsd.org>
initial attempt at a kdemultimedia package.
enough patches to make everything compiles, lots of patches to
rename the internal mpeglib and workman libraries to something that
doesn't conflict with the rest of the ports tree.

kscd works, noatun and skins as well.

mpeglib seems to have some threads issue yet, but it compiles.