Written by: Marc Espie <espie@openbsd.org>
let build of dependencies that are not 'install' or 'package' happen
inside a subdirectory of the working directory, instead of elsewhere.
Allows packages BULK=Yes to work correctly, and increase reproducibility
of builds.

tested by naddy@ through several incarnations.

Ports that use patch/configure/build dependencies must be adjusted for this.
Pleasingly enough, the adjusted version is often simpler.

The change uses an internal _MASTER variable to recall the list of ports
built during dependencies, for less confusion.