Written by: Peter Valchev <pvalchev@openbsd.org>
sync with in-tree binutils:
date: 2002/03/19 02:02:57;  author: pvalchev;  state: Exp;  lines: +27 -5
When one shared library requires another, it will try to locate the
required shared library and include it in the link if it's not specified
explicitly.  This change makes ld look at the -L arguments specified,
and build search path.  This avoids the need for -rpath-link, and
somewhat resembles the a.out linker behaviour.  However behaviour of
-rpath and -rpath-link is still preserved as a GNU extension.