Written by: Peter Valchev <pvalchev@openbsd.org>
Add -L${X11BASE}/lib to sdl-config in when X11 is used, as SDL depends on
libraries there and needs to know the path.  this means sdl-config was
broken for a few days but none of the ports depended on this.
Also bump package revision to p1 for the last couple of fixes to the port.