Written by: Peter Stromberg <wilfried@openbsd.org>
Update to Net-DNS-0.14

This is an interum release, intended to return Net::DNS to CPAN,
and encourage new devopement.  Version 2.0 is not far behind with
major changes including support for windows platforms, TSIG and
Net::DNS::Nameserver (a very simple nameserver).

Fixed the "defined(@array) is deprecated" message in DNS.pm under
perl 5.6.0 and greater.

Fixed a bug in the Net::DNS::Resolver->nextid method to ensure that
the query ID remains 16 bits.