Written by: Damien Couderc <couderc@openbsd.org>
Initial import of editors/beaver
--
Beaver is an Early AdVanced EditoR, for Linux and other Unices (and even
Windows); in other words, it's a text editor that is intended to be
light-weight, but full of useful features for programming, from editing of web
sites to C coding. It is based upon the GTK+ toolkit, supports many languages
through config files and offers functions such as automatic indentation and
completion or syntax highlighting. In fact, it is 100% compatible with
Ultraedit's wordfile.txt files, and has its own mini macro language.