Written by: Anil Madhavapeddy <avsm@openbsd.org>
Import php4-core-4.2.1

Installs the barebones php4 with only the gettext, iconv and recode
modules compiled in.

All of the other modules have to be installed as shared modules on
top of this.

In addition to the Apache module, this package also includes a php
command-line binary which can be used in shell scripts.  The binary
uses the same /var/www/conf/php.ini file as the Apache module.

There is some non-i386 breakage at the moment (notably macppc).

Work by wilfried@ and me.  espie@ ok