Written by: Shell Hung <shell@openbsd.org>
Initial import of Font-AFM-1.18 (Used by p5-Mail-Box)

perl interface to Adobe Font Metrics files
---
This module implements the Font::AFM class. Objects of this class
are initialised from an AFM-file and allows you to obtain information
about the font and the metrics of the various glyphs in the font.