Written by: Kevin Lo <kevlo@openbsd.org>
Initial import of p5-IO-Interface 0.97.

IO::Interface adds object-methods to IO::Socket objects to allow
them to get and set operational characteristics of network interface
cards, such as IP addresses, net masks, and so forth.  It is useful
for identifying runtime characteristics of cards, such as broadcast
addresses, and finding interfaces that satisfy certain criteria,
such as the ability to multicast.