Written by: Marc Espie <espie@openbsd.org>
simple update of ogle-gui (compiles like a charm).
Turns out the audio issues of ogle are not mpeg-related at all,
but to ogle getting confused as to what audio stream it should be
decoding.

So, if you get silence, try using ogle_gui audio button instead.
The next version of ogle may fix that, now that the problem is
identified.